First Pass - Replace hotkeyPress key handlers.
As UI buttons are created each frame, they're dropped in to
`KDButtonsCache[]`.  This was done using "free-standing" untyped
objects.  The array was scanned each frame to look for mouse clicks and
key presses, and invoke the corresponding callback function when
appropriate.

It also became apparent that having to populate the callback function
for every key definition was going to become annoying boilerplate,
especially when the button already has a callback function field.

Create `ButtonCacheEntry` type, so we have an actual type to work with
and check; `KDButtonsCache[]` is now an array of this type.

Make `KDKeystrokeDef.func` optional (don't worry, read on).

Create convenience function `KDRegisterKeyDefs()`, which registers a
`KDKeystrokeDef` into the `KDKeystrokeCache[]`.  Further, if the `func`
callback field is empty, this will populate it with `commonFunc`.

Create convenience function `KDRegisterKeyDefsCompat()`, which takes a
`Button` object and converts the `hotkeyPress` into a `KDKeystrokeDef`,
then clears the `hotkeyPress` property (this prevents it from being
handled by the legacy key handler).

Create convenience function `KDRegisterButton()`, which lets us
seamlessly replace the direct array assignments to `KDButtonsCache[]`.

Also:

Rearrange logic in `KDMouseWheel()` to be clearer.

Add new "static" variable to `KDDrawInventoryContainer()` to hold scroll
increment for inventory filter list; add logic to apply it.

All told, the idea is to change out the key event handling from
underneath all the buttons.

Unfortunately, it doesn't completely work, as I b0rk3d handling of
modifier keys.  I'm going to have to reconsider how to deal with Shift,
Ctrl, Alt, and Meta keys, as the game only uses them as simple buttons.