Fix handling of Inventory vs. Container list scrolling.
I noted a bug when browsing the Warden's Chest -- where all the articles
used by the designated Warden of The Summit to recapture prisoners are
stored.  When clicking on its scroll gadgets, the other (inventory) list
was scrolled.

The state needed to scroll what was displayed was improperly captured.

Capture the proper scrolling state, depending on whether we're
displaying the player's inventory, or some other list.  Apply said state
on the next pass through the function.