Add Home/End hotkey support for Summit collection.
This turned out to be more complicated than originally thought.  The
main loop goes through the entire `KDGameData.CollectionSorted` array,
but only a subset of those entries are eligible for listing in the main
collection.  Thus, attempting to adjust the scroll index to the "end"
must take this in to account, as `KDGameData.CollectionSorted.length` is
almost always too long.

Move addition of scroll up/down buttons to the bottom of
`KDDrawCollectionInventory()`.  In the main loop, count the number of
eligible entries for display, then use this value to calculate the
correct value for the index.